refactoring code to different modules for all market functions moved into market.py and the logging and output into the results folder
